core: prevent NPE when vm disappears from the host
It is possible that from the monitoring point-of-view a VM disappears
from the host. For instance, when powering-off a VM it suddenly stops
being reported by the host (power-off calls 'destroy' so the VM is
dropped from the list of running VMs on the host). In that case, the
code that prepares the VMs for devices monitoring may lead to NPE.
This patch prevents it by adding a null-check (obviously, the devices
of such VMs should not be monitored at that point).

Change-Id: I541bc763c99187556a5d4d76cb46df7045f8bab8
Signed-off-by: Arik Hadas <ahadas@redhat.com>