frontend: fix ClassCastException in StorageModel
StorageModel::updateDiscardAfterDelete used to set the
Discard After Delete (DAD) entity to false if it's not a block storage
domain. This caused DAD's entity changed event to be raised, which
should happen only for block domains.

This patch removes the redundant call to setEntity on DAD, as it's
already being set to false when switching back to a block domain.

Change-Id: Ie31f4e195569b2ce931f82d856e364e274825349
Bug-Url: https://bugzilla.redhat.com/1451246
Signed-off-by: Idan Shaby <ishaby@redhat.com>