resatpi: Mocking backend exceptions in tests
In the backend, once a VdcQueryRetrunValue is generated, it can no
longer throw an exception when it's accessed, so mocking
queryResult.getExceptionString() to throw an exception is artificial,
and doesn't represent any real world use case.

This patch moves the mocking of throwing a [runtime] Exception to the
backend.runQuery call, which represents a query that actually fails
catastrophically.

Change-Id: Ibd9d1f5555c5a357f9947f727a7b6f1c5ea568bf
Signed-off-by: Allon Mureinik <amureini@redhat.com>