core: SyncLunsInfoForBlockStorageDomain - domain md devices
When a domain is created it has a vg metadata device and a metadata lv
which may be created on multiple devices. On a regular basis, those
devices should be never changing.

However, in some user envs in case of disaster the block sd may be
resotored manually in a way that will change its metadata devices.

The SyncLunsInfoForBlockStorageDomain is executed when a block domain
moves to Active status, it synchronizes the domain luns info persisted
in the db with the actual info retrieved from the storage - this patch
adds to it the code that'll update the info on the domain metadata
devices as well as it may changed.

Change-Id: Id65dad542fee6f1a55625f5e5730b112f8149e0e
Bug-Url: https://bugzilla.redhat.com/1452984
Signed-off-by: Liron Aravot <laravot@redhat.com>