webadmin: Properly handle empty disk selection
Commit 0af332 that was issued to fix bug #1452790 solved the UI
freezing problem, but missed the distinction between passing an empty
list CreateAllSnapshotsFromVmParameters (no disks should be included
in the snapshot) and a null list (all the disks of the VM should be
included in the snapshot).

This patch fixes that mistake.

Change-Id: Ie69b1f2dbef9c8386c0ae8363145ec3d0c90b74c
Bug-Url: https://bugzilla.redhat.com/1452790
Signed-off-by: Allon Mureinik <amureini@redhat.com>