core: CommandsFactory generics
CommandsFactory#commandsCache holds instances of both command and
query classes. Its generic specification is just plain wrong, and the
code only works due to the runtime type erasure.

Any class returned from this class' methods is either used regardless
of its generic specification, or instantiated and then cast to the
right type, making the generic specification somewhat of a mute
point.

This patch replaces it with a straight-forward Class<?> in order to
both avoid the confusion of the first paragraph and set up the
following patches.

Change-Id: I79d387829d524d8285dddad370a54e427ca3ff5e
Signed-off-by: Allon Mureinik <amureini@redhat.com>