core: CommandsFactory computeIfAbsent
Use the JDK's Map#computeIfAbsent method instead of reinventing the
wheel.

Besides making the code a tad more elegant, in the edge case where
two threads try to load the same command for the first time, this
variation of code should perform a tad faster, as the findClass
lambda is guaranteed to execute only once.

Change-Id: I85d7c02f3c984919b1a73bd913e0b57f992bc3b8
Signed-off-by: Allon Mureinik <amureini@redhat.com>