backend: introduce SyncStorageDomainsLunsCommand
This patch introduces a new command to synchronize all active block
storage domains in a given data center.

Previously, when synchronizing more than one storage domain (when
upgrading a data center or when activating a host where the other hosts
in the data center are not active), the engine used to call
SyncLunsInfoForBlockStorageDomainCommand for each storage domain.
The problem was that SyncLunsInfoForBlockStorageDomainCommand used to
call GetVgInfoVDSCommand, so for each storage domain we did a full
roundtrip to vdsm and back to the engine.

With this patch, SyncStorageDomainsLunsCommand calls GetDeviceList only
once, and uses that information to update all the domains in the data
center. Therefore, when synchronizing n domains, the engine should do
it ~ n times faster than it did before.

Change-Id: I744ae190397e00856ed67b99046efd1b1bb26054
Related-To:https://bugzilla.redhat.com/1404389
Related-To:https://bugzilla.redhat.com/1455547
Related-To:https://bugzilla.redhat.com/1455548
Signed-off-by: Idan Shaby <ishaby@redhat.com>