core: Standardize queries constructors
Currently, queries (i.e., classes under the QueriesCommandBase
hierarchy) may have one of two constructors - either a constructor
with some argument that extends VdcQueryParametersBase, or a
constructor with two arguments - one that extends
VdcQueryParametersBase and an EngineContext.

This uncertainty may lead to all sorts of hard-to-find bugs where a
context is either passed to Backend#runQuery and then silently
ignored, or worse, a query could not be executed from inside a
command if it doesn't have the right constructor.

This patch streamlines all the queries to have a single constructor,
with a parameters class and a context. As always, passing a null
context will cause a default context to be created (by the
QueriesCommandBase(VdcQueryParametersBase, EngineContext)
constructor.

Change-Id: Ia790a35b75093210c8559590f4878ac42414e8bb
Signed-off-by: Allon Mureinik <amureini@redhat.com>