core: SyncLunsInfoForBlockStorageDomain - metadata devices refresh
This patch changes the SyncLunsInfoForBlockStorageDomain command to skip
executing the metadata devices refresh if no vdsId was passed to the
command.

refreshMetadataDevicesInfo() is executes
BlockStorageDomainHelper.fillMetadataDevicesInfo() which currently
doesn't use an api that support randomizing a host for execution.
Currently all the flows the needs that functionallity do pass the vds
id, while the command should obviously not be aware to the parameters
passed by it's callers -  as this patch is intended to be backported
to the 4.1 branch the change here is minimal - skips the execution if
no vds id was passed.

In the master branch fillMetadataDevicesInfo() will be updated to
supported randomizing a host which will enable us to remove that check.
If we want to skip the metadata devices refresh it should be done
using a specific parameter.

Change-Id: I35bdcbe4d5ff55e2fbe39ca4ab28b05b4802454b
Bug-Url: https://bugzilla.redhat.com/1452984
Signed-off-by: Liron Aravot <laravot@redhat.com>