core: Fix permissions subtab of users
Previously the permissions sub-tab of user tab displayed permissions
of the user, which has assigned on specific object directly or
in-directly(inherited). But if the user was part of the group, those
permissions wasn't displayed. We displayed the permissions of the
groups of the currently logged-in user instead.

This patch fixes that, so we now display the permissions of the groups
of the user, which is selected in 'Users' tab instead of the user which
is currently logged-in.

Change-Id: Icbc1ee3b865349384b95a1d5f09880a448712359
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1455011
Signed-off-by: Ondra Machacek <omachace@redhat.com>