core: ProcessOvfUpdateForStoragePoolCommand - db updates
ProcessOvfUpdateForStoragePoolCommand.performOvfUpdate() iterates over
the ovf data and updates it to the db, the max limit is for items per
update is OVF_MAX_ITEMS_PER_SQL_STATEMENT.
The lists are being initiated after the first loop, in case more than
one iteration not all the updates will be performed.
This case was never actually reached because the number of ids processed
is always lower or equals to the OVF_MAX_ITEMS_PER_SQL_STATEMENT - This
patch fixes that wrong placement of the initiation for correctness.

Change-Id: I968c9cd789a52e66a0c91d33629808e368bc29d1
Signed-off-by: Liron Aravot <laravot@redhat.com>