backend: override applyLockProperties in SyncLUNs
The method applyLockProperties was not overridden by
SyncLunsInfoForBlockStorageDomainCommand and thus the exclusive lock
with the group SYNC_LUNS on the storage domain's id was never taken
before running this command.

Since we don't want two SyncLunsInfoForBlockStorageDomain commands to
run concurrently for the same block storage domain, this patch overrides
this method so that lock will be taken.

Change-Id: Ica5aad22abd354c94110ab14d371be6d63ef43b1
Signed-off-by: Idan Shaby <ishaby@redhat.com>