core: NumberOfMonitorsConstraint null checks
The ValidNumOfMonitors config value should always be present. If it
isn't, there's a problem with the installation, and the code should
error out instead of silently failing the validation.

This patch removes the wrong null check on the config value's result.
With this check eliminated, the code can be slightly optimized to
avoid retrieving the config value if the passed value was null.

Change-Id: Ib5a74c917336cdc0d60b7559abb20710f8d123e9
Signed-off-by: Allon Mureinik <amureini@redhat.com>