search: fix tags/values with underscore behaviour
This patch fixes a regression caused by [1] that attempts to resolve [2]
The fix caused tags including an "_" attached to Host or VM not to work.
The problem was that the [1] patch put "\" before "_" not only for a
condition with LIKE/ILIKE but also for "=" condition which cause the
search to return empty results.

The problem is that we have generated queries with ILIKE/LIKE in two forms :
1) <field> LIKE/ILIKE <value>
2) <value> LIKE/ILIKE ANY (string_to_array ......

[1] https://gerrit.ovirt.org/#/c/69926
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1409849

Change-Id: I3d8e2075f99ce2600e64bd996280c68feb8f3fa0
Bug-Url : https://bugzilla.redhat.com/show_bug.cgi?id=1446525
Signed-off-by: emesika <emesika@redhat.com>