core: change LSM lock scope to execution
Since the memory lock on the disk is not needed for
the entire command and only needed until the DB lock
is acquired, the scope can be changed to execution.

Having the lock in the Command scope also causes a
race condition if an operation runs after LSM tasks
are done but the engine lock was not released yet.

Change-Id: Ic2edf9ac086ed3b647e68379f3507443308e3a0c
Bug-Url: https://bugzilla.redhat.com/1456504
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>