core: Refactor of setting disk attributes.
Engine updates the new disk image with updated size and actual size
attributes based on the downloaded image retrieved from Glance.
The engine reads the image using inputstream and therefore
makes it hard to maintain for extra fields to update.

The proposed patch refactors the update of the disk image fields, to
be done inside the parsing method, making diskImage a parameter in
the method signature.

Change-Id: Iea7f2c6b566d9436d9799c56e86901e956ec49ae
Bug-Url: https://bugzilla.redhat.com/1456432
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>