backend,frontend: add hugepages reporting and display
The huge pages are reported from VDSM in the vds statistics.

It is a map structured like this:
    "hugepages": {
        "1048576": {
            "resv_hugepages": 0,
            "free_hugepages": 0,
            "nr_overcommit_hugepages": 0,
            "surplus_hugepages": 0,
            "vm.free_hugepages": 0,
            "nr_hugepages": 10,
            "nr_hugepages_mempolicy": 10
        },
        "2048": {
            "resv_hugepages": 0,
            "free_hugepages": 0,
            "nr_overcommit_hugepages": 0,
            "surplus_hugepages": 0,
            "vm.free_hugepages": 0,
            "nr_hugepages": 0,
            "nr_hugepages_mempolicy": 0
        }
    }

It is a map where the key is the size of the huge pages and the value are the
details.

This structure is than transformed on engine side into a json list with the
following structure:
[{"sizeKB":1048576,"amount":4},{"sizeKB":2048,"amount":0}]
e.g. size of the huge page in KB -> amount of free huge pages available for VMs
(vm.free_hugepages) of this size.

This json is than stored into the vds_statistics table and later shown in the
general subtab of the host main tab.

Change-Id: I249ad336bdeeda1e58001222049c2e824c6634a3
Signed-off-by: Michal Skrivanek <michal.skrivanek@redhat.com>