frontend: set DAD to false when entity is unavailable
Patch a2e073845dc5485b36a6bad00c1a8e63c238342b indeed fixed a UI bug as
requested, but it also made it impossible to:
1. Navigate to new block storage domain window.
2. Set DAD to true.
3. Navigate to file domain and create one.

Since the last time DAD was available it was set to true, and since only
the availability of DAD is set to false, the value remains true and the
frontend tries to create a file domain with DAD = true ,which is
impossible, and the user gets an error message.

This patch sets DAD to false whenever it is unavailable.

Change-Id: I7dfaa3e5e8ab083d2c0e4f3409e24a1b5111487b
Bug-Url: https://bugzilla.redhat.com/1458568
Signed-off-by: Idan Shaby <ishaby@redhat.com>