core: Use disk id instead of volume id in OvfUtilsTest
This fix is related to test of unregistered disks.
The engine filters out disks which are attached to VMs to know
which are unregistered disks and which are not.
Since the filtering is being done with the disk id the test should
be changed to use the disk id in the OVF instead of the volume id.

Change-Id: Ib5d1f613a6ead694e93f8a29cda53b0f24d6de03
Bug-Url: https://bugzilla.redhat.com/1443641
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>