core: Free pooled VM lock independently from context locks
When a non-prestarted VM is allocated from VM Pool, it is locked early
in AttachUserToVmFromPoolAndRunCommand flow, before other locks of the
command are taken. It is possible that it is also freed before that
moment, so context.getLock() != null condition will be false, but the VM
lock is already taken and must be freed.

Change-Id: I628946d781ecbe280562efe7cf7e3a3c84801f39
Bug-Url: https://bugzilla.redhat.com/1433924
Bug-Url: https://bugzilla.redhat.com/1432105
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>