core: change order of validations in RemoveStorageDomainCommand
Currently when attempting to remove a storage domain, part of the
validation process attempts to mount a storage domain to check if
it's attached to a DC and then disconnects. If the storage is already
attached, disconnecting from it may trigger a master domain
reconstruction (if it's the master) and it shouldn't do that.

The patch moves the isStorageDomainAttached check to the end of the
validation since we can fail earlier by querying the database and
discovering the domain is already attached without triggering the
reconstruction.

Change-Id: Id13fa174e79441b6c0641b7b7190efe16f99fbfc
Bug-Url: https://bugzilla.redhat.com/1444120
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>