Add a new 'High Performance' VM type to UI
This patch adds a new "optimized for" field type to VM dialog in UI,
in addition to the Desktop/Server types.
This new "high performance" type will be used for providing high performance
VMs with relevant configuration and validations.

This patch is adding the new 'high performance' option to the UI
with relevant new icons, tooltips and basic validations.

On top of this patch, we will add the configuration settings
relevant for this new VM type.

Change-Id: Ic4d5e6ad1c3b1f582a1646ecf2f3b7fbdaeeb3c9
Bug-Url: https://bugzilla.redhat.com/1457239
Signed-off-by: Sharon Gratch <sgratch@redhat.com>