userportal, webadmin: work around GWT-P memory leak
We encountered a fairly large memory leak in the latest
version of GWT-P. All GWT-P popups have a resize handler
attached to them, but this handler lives at window scope
and never lets go. Therefore, every popup is completely
leaked when closed.

However, luckily, the method in GWT-P that sets up this
handler has a hook/template method for disabling the addition
of the handler:
    if (repositionOnWindowResize()) {
         Window.addResizeHandler(new ResizeHandler() ...

We simply overload repositionOnWindowResize() to always
return false and the leak is worked around.

See: https://github.com/ArcBees/GWTP/issues/823

Change-Id: I4e6d4da68e3f50f5863f4e33e62581b97e36dd2d
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>