webadmin: ListModelObjectCellTable ArrayList ctor
This patch creates an ArrayList from the getSelectedSet() by using
the overloaded consturctor instead of looping over getSelectedSet()
and calling List#add for each element.
At worst, this syntax is just cleaner and easier to understand. At
best, it may offer a slightly improved performance, as the ArrayList
has prior knowledge of the number of elements being added to it, and
can thus avoid resizes.

Change-Id: I0ab0b97006e95007e8495a7c017b563bcc148de8
Signed-off-by: Allon Mureinik <amureini@redhat.com>