core: Remove failure reasons from EngineLock#toString
EngineLock#toString prints the validation failure that should be used
if the lock cannot be acquired. This leads to INFO messages about
successful lock acquisitions to seem, at first glance, as error
messages, which is confusing (see, e.g., the attached bug).

This patch suggests improving the logs by NOT printing the validation
messages, which will only be printed (and only be useful) if the
engine fails to acquire the lock.

Change-Id: I6ffc9107f67ffe5896e1a61b5a4511b83168c26e
Bug-Url: https://bugzilla.redhat.com/1454843
Signed-off-by: Allon Mureinik <amureini@redhat.com>