backend: fix log with empty removed luns list
When SyncLunsInfoForBlockStorageDomain removes old luns from the db, it
logs "Removed LUNs, IDs '<luns_list>'".
When there are no luns to remove, the following useless log was still
printed: "Removed LUNs, IDs ''".

This patch fixes this bug so no log will be printed if there are no luns
to remove.

Change-Id: I072af233a3201980f69e73806fb068e6f1c39c1d
Bug-Url: https://bugzilla.redhat.com/1460512
Signed-off-by: Idan Shaby <ishaby@redhat.com>