core: CommandBase#getActionReturnValue return type
VdcReturnValueBase#getActionReturnValue returns a generic type T,
which Java 8 allows us to infer.
CommandBase#getActionReturnValue just calls that method, but returns
an Object, forcing the callers to downcast it.

This patch fixes CommandBase#getActionReturnValue return type so it
can also be inferred in order to clean up the code a bit.

Change-Id: I0cf884a78732bf465d90f4e42a1e4d68f762da02
Signed-off-by: Allon Mureinik <amureini@redhat.com>