packaging: Add bug to PostgreSQL JDBC driver issue
In patch I19b76230d5551b718349dbe14f7a01c4349e23b3 we have removed
dependency on postgresql-jdbc on Fedora due to
https://github.com/pgjdbc/pgjdbc/issues/838 tracked by oVirt bug
https://bugzilla.redhat.com/show_bug.cgi?id=1459502 but we haven't
mentioned the bug in spec. So this patch adds the bug to spec file
comment so we don't forget to return postgresql-jdbc dependency once
Fedora package is fixed.

Change-Id: If4394cd8de7c844c19991f5f6c6268249f0e711e
Signed-off-by: Martin Perina <mperina@redhat.com>