webadmin: Add ErrorMessagesTest to frontend module
Commit 25603e [rightfully!] removed ErrorMessagesTest from the
webadmin module, as it no longer has an AppErrors.properties file,
after it was removed in 6d4008.

That patch, however, neglected to add an ErrorMessagesTest to the new
frontend location.

This patch adds back the missing test and removes the redundant key in
the properties file found by the said test.

Change-Id: I570eb1d7f48a4b53e5408e5d670f6d4dc6b607d9
Signed-off-by: Allon Mureinik <amureini@redhat.com>