backend: make vdsId mandatory in SyncLunsInfo cmd
Previously, SyncLunsInfoForBlockStorageDomainCommand could optionally
get a VDS ID as a parameter.
When passed, it was used to call GetVGInfo, and when it wasn't, a random
and active VDS in the storage pool was chosen to do that.

The problem is that refreshMetadataDevicesInfoIfNeeded counts on the
existence of this VDS ID in the parameters. When it's null, the method
logs that the metadata devices refresh was skipped.
The method's comments say that "Currently all the flows...pass the vds
id...", but it's not true when upgrading the storage pool (was
introduced by ca8f351f1b4fef1f134eae0790bc07b1eab32aa4).

Since comments don't run, and since we don't want to skip the metadata
refresh, this patch adds the VDS ID parameter to all the constructors of
SyncLunsInfoForBlockStorageDomainCommand to make it mandatory.
Thus it makes sure that an active VDS ID is passed to it when
called from SyncStorageDomainsLunsCommand, which is used on the upgrade
storage pool flow.

Change-Id: I473e7d54570afcf39fc31d578a60fa9499f97920
Bug-Url: https://bugzilla.redhat.com/1465839
Signed-off-by: Idan Shaby <ishaby@redhat.com>