core: Allow to migrate a paused VM with Post-copy policy
Libvirt does not allow to migrate a VM in Paused state with post-copy
migration. If migration policy of a cluster contains a "postcopy"
action, every attempt to migrate a paused VM fails because libvirt
returns an error, not allowing migration to start. On the other side,
the current Post-copy migration policy uses post-copy only as last
resort, when a regular migration stalls. It is unlikely that migration
will be stalling when migrating a paused VM, so it is safe to use a
Post-copy policy with it.

Libvirt bug have been open for this issue:
https://bugzilla.redhat.com/show_bug.cgi?id=1464792

This patch introduces a workaround: MigrateVmCommand will filter a
"postcopy" action out of the effective migration policy, if the VM is in
Paused state. This helps to avoid libvirt error and allows to migrate
such a VM in a cluster with Post-copy migration policy.

Change-Id: I5fcaf057f3d3443d5d0dd30e542a8e159acc5bd4
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>