core: toward deprecation of the speed of vm interface
Today, VDSM reports the speed of a VM network interface. It just determines
that according to the model of the interface (e1000/virtio/...) [1]. This
is redundant since the engine already knows the model of the interface,
so it can figure out the speed on its own.

This patch introduces a preparation for deprecating this field. The reason
for doing it now is that the speed is not reported by collectd and we do
not wish to add it to collectd.

[1] https://github.com/oVirt/vdsm/blob/master/lib/vdsm/virt/vmstats.py#L286

Change-Id: Iafb14eb481d099ec89ed8a8b038b9d4db02fb63b
Signed-off-by: Arik Hadas <ahadas@redhat.com>