webadmin: Fix handling of empty network name
If the user manually changes the name of the default vNIC profile,
it is not synchronised to the network name anymore.

This synchronization can be broken, by synchronize the name of the
default vNIC profile name to an empty network name, which is
represented as 'null' value in the entity.

This changes fixes the behavior by not stopping the synchronization,
if the default vNIC profile name is 'null'.

Change-Id: I78fa9b356e51f49e833d12cbd3bdc36799099938
Bug-Url: https://bugzilla.redhat.com/1459003
Bug-Url: https://bugzilla.redhat.com/1463568
Signed-off-by: Dominik Holler <dholler@redhat.com>