webadmin: Introduce OvirtSelectionModel
Use new higher-order component, OvirtSelectionModel, which
satisfies the given interface (GWT SelectionModel) through
delegation to the appropriate implementation. This makes the
code (e.g. ListModel) cleaner and easier to understand.

Mark ListModel#getSelectionModel method as protected because
it's meant to be called by constructor (only) with subclasses
having the ability to override the default value.

Change-Id: I4cc0e2d5f12ee1a6281828242ebb60918591f8a3
Signed-off-by: Vojtech Szocs <vszocs@redhat.com>