webamdin: Standardize AbstractVmPopupWidget#DataCenterWithClusterComparator
DataCenterWithClusterComparator is a boilerplate implementation for
sorting DataCenterWithCluster objects by the DC name and then by the
cluster name.

This patch standardizes the implementation by removing the
boilerplate code and replacing it with a chained
Comparator.comapring(...).thenComapring call.

The original code seems to have several slight bugs which do not seem
to be the intended behavior based on the class' comment ("Comparator
that sorts on data center name first, and then cluster name. Ignoring
case."), and may lead to inconsistent sorting:
- Two objects with a null DC name will be considered equivalent,
  regardless of the cluster name. This seems inconsistent with the
  behavior of sorting two objects with the same DC name by the
  cluster name.
- Moreover, for two objects have a non-null DC name, if they have
  equal (*case sensitive*) DC names, they are sorted by the cluster
  name, but if the DC names are not equal, they are sorted by it in a
  *case insensitive* order.

The revised code takes a cleaner, more obvious approach:
- Sort by the DC names, in a case insensitive order, with the nulls
  last.
- Break ties by sorting by cluster names, in a case insensitive
  order, again with nulls last.

Change-Id: I6e5a06a8af094b3693228c29ed87182b4ae3183a
Signed-off-by: Allon Mureinik <amureini@redhat.com>