core: add vm boot time
There are few approaches for calculating the elapsed time of a VM.
One approach, that is used today, is to aggregate the time the VM
was running. This approach has some drawbacks.
An alternative approach is to aggregate the time the VM was down
and then to compute the elapsed time according to the current time,
the overall downtime, and the boot time (that is, the time the VM
was started and not restoring an existing memory state).
Since we are going with the latter approach, this patch adds the
VM boot time.

Change-Id: If938ab3c6def74c39ce8085c21aa80c88a8eb4cb
Signed-off-by: Arik Hadas <ahadas@redhat.com>