core: Add audit entry with cold merge info
In commit 2420937 (cold-merge: Update step message to include the host
name), we added merge info, mainly the host running the merge job, to
the step message. In this patch, we introduce a mechanism to add an
audit log entry.
In order to add an audit log entry:
1. An audit log type with same step enum name must be added to
   AuditLogType.java
2. A message must be added to AuditLogMessages.properties. The arguments of
   the message must fit those in the message of the step enum. Usually,
   the same message is used

For example, for cold merge we previously added MERGE_SNAPSHOTS_ON_HOST
in StepEnum. Now, we add MERGE_SNAPSHOTS_ON_HOST to AuditLogType.java and
add a corresponding message (actually same message created for the
step enum) in AuditLogMessages.properties.

Change-Id: I685e9d833e67aa9ef0ea9680cdbacc1cc318eef8
Bug-Url: https://bugzilla.redhat.com/1464795
Signed-off-by: Ala Hino <ahino@redhat.com>