backend: DAD's default value criteria
When creating a block storage domain without stating the value of
Discard After Delete (DAD), its default value is calculated from the
data center's version of the host that added that domain.

To reduce the coupling between the host that performs the operation and
DAD's default value, this patch changes the way it is done to be
according to the storage format - the data center version is calculated
from the storage format, and the default value of DAD is calculated from
the data center's version.

Note that not specifying the storage format causes the storage format to
be calculated from the data center version of the host that added that
storage domain. So creating a storage domain with a host that is
attached to a data center with version 4.0 will cause DAD to be false by
default, while version 4.1 will cause it to be true by default.

Also, this patch removes a validation part of DAD in
AddSANStorageDomainCommand, since it used to check if DAD was legal
according to the data center version of the host that added that
domain. There's no reason to validate if DAD is legal in the data center
since AddSANStorageDomainCommand only creates an unattached storage
domain. This should be (and indeed is) checked in
AttachStorageDomainToPoolCommand.

Change-Id: I9a6a07878482bec0fe9c2c552c284bd95630c0f4
Bug-Url: https://bugzilla.redhat.com/1463083
Signed-off-by: Idan Shaby <ishaby@redhat.com>