core: Fix potential OvfParserTest NPE
OvfParser#utcDateStringToLocalDate may potentially return null (even
though this is not possible in this test due to the hard-coded
argument that's passed to it), causing static analysis tools such as
Coverity or even IntelliJ IDEA to issue a warning on this test.

This patch flips the test around and constructs a new Date object
from the expected timestamp instead of extracting the timestamp from
the parsed Date object. This change not only silences the
aforementioned static analysis tools, but also makes the test a tad
more resilient - if utcDateStringToLocalDate would ever be broken in
the future in a way where it returns null for this argument the test
would now fail cleanly instead of erroring out with a
NullPointerException.

This patch addresses Coverity CID 1376661.

Change-Id: If23a59cd5f85085bd6c965ab653238c351e9d8a1
Signed-off-by: Allon Mureinik <amureini@redhat.com>