core: creation date in the active snapshot on preview snapshot
The creation date on a VM snapshot is supposed to indicate the time
that the snapshot was taken. Therefore, when previewing a snapshot
the creation date of the active snapshot needs to be set to the
creation time of the snapshot being previewed. However, today the
creation date is always set to the time the snapshot entity was
created, also for active snapshot which is wrong and thus changed
in this patch.

Change-Id: I8e3ae76cbba3b96059aeb9d8b74ece66c82a223e
Signed-off-by: Arik Hadas <ahadas@redhat.com>