engine: Detach RemoveImage from MoveDisk command
At the end of the MoveImageGroupCommand, a RemoveImage command
is called to clean the source disk.

In the case the RemoveImage is taking long to finish, the Job
of the MoveImageGroupCommand could be cleared meanwhile and when
the RemoveImage command will try to update its step it will fail to
find the Job entry causing an SQL exception.

In order to avoid this situation, the RemoveImage command will
be executed in a new context.
Note that its execution will not be reported in the UI.

Change-Id: I86deaf1f3d716e28e76236c73f13c830b3c7338d
Bug-Url: https://bugzilla.redhat.com/1464348
Signed-off-by: Fred Rolland <frolland@redhat.com>