webadmin: replace blank white screen with 'Loading' spinner
Previously, after logging in, there was a 5 to 10
second white screen pause while the initial GWT
JS script downloaded, JIT compiled, and rendered the
GWT UI.

This patch adds a PatternFly loading spinner to the GWT
host page, which is the root page of the webadmin
(i.e. the white screen == the host page).

Now, while GWT JS selector script is doing its work,
this 'loading page' is shown.

We still want to reduce the size of the initial GWT
download. This is a good intermediate fix.

Change-Id: Ie5ef8377a1e555316b94ddf6d6310408e11c5a01
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>