core: Detach compensation on startup from CommandBase
The compensation on startup was designed to return the DB to its consistent
state. The compensation performed on startup is not related to the
commands themselves, but for the business entities snapshots.

This patch extracts the compensation handling from the commands, so they
could be used without a command instance.

By doing so we allow the following:
1. Perform compensation without instantiating a command and without to
require all of the command's injected bean and dependencies.

2. Allow to place the compensation logic as soon as possible when the
system starts up.

3. Eliminate the need to declare the dedicated c'tor in each command's
class that should be compensated (c'tor that comply to CommandBase(Guid)).

A following patch should remove all of the CommandBase(Guid) signatures
from the code for all of the classes, as it won't be use any more.

Move initializing all command related services to InitBackendServicesOnStartupBean
after initializing VmPropertiesUtils, DevicePropertiesUtils and
HostDeviceManager. This should avoid the circular dependency when
CommandRepository.handleUnmanagedCommands tries to require locks
for unmanaged commands on startup.

Change-Id: I5d8b2f1ddd469c13684057cc2a5af440d6f5f8f0
Bug-Url: https://bugzilla.redhat.com/1525569
Signed-off-by: Moti Asayag <masayag@redhat.com>