backend: improve block domains sync on host activation
When activating the first host in the data center, all the storage
domains in the data center are activated.
Today, the engine calls SyncLunsInfoForBlockStorageDomain for each
storage domain in a new thread using the same host.
This patch improves this behavior by synchronizing all the domains
together using SyncStorageDomainsLuns and thus performing one call to
vdsm (instead one per domain) in one thread.

Change-Id: I6539a2e96ac885f828d93276bf939931bdfc372c
Bug-Url: https://bugzilla.redhat.com/1467794
Signed-off-by: Idan Shaby <ishaby@redhat.com>