core: Refine live merge validation
Currently, there is a validation that disable running concurrent live
merge even for snapshots that created on different disks. In this patch,
validation is refined to disable executing concurrent live merge
operations only if the snapshots contain same disk(s).

Note that this new validation is only used for live merge. There are
other operations (add Vm from snapshot, create Vm template, migrate Vm
and remove Vm) that will still use the previous validation. In case we
decide we want to refine the validation for those operations, this will
be done in separate patches.

Change-Id: I984b4ccf10993e982363e8b35871e97fed57e56e
Bug-Url: https://bugzilla.redhat.com/1465539
Signed-off-by: Ala Hino <ahino@redhat.com>