backend: DAD initialization in AddExistingBlockStorageDomain
AddExistingBlockStorageDomain deals with block domains, and thus should
have its own validation and default value for DAD, similar to
AddSANStorageDomain (and not inherit file domains implementation from
AddStorageDomainCommon).

Since the AddStorageDomainCommand hierarchy is not structured by the
storage type, this patch adds a default implementation for
validateDiscardAfterDeleteLegal in AddStorageDomainCommand and overrides
it in commands that handles block domains.

Change-Id: I8e506a92c0534118805de6c9c7adbaa8df5eb9b1
Bug-Url: https://bugzilla.redhat.com/1467781
Signed-off-by: Idan Shaby <ishaby@redhat.com>